+++ src/Stripe.net/Services/SubscriptionSchedules/SubscriptionScheduleAmendOptions.cs
@@ -13,6 +13,12 @@ namespace Stripe
         public List<SubscriptionScheduleAmendmentOptions> Amendments { get; set; }
 
         /// <summary>
+        /// Provide any time periods to bill in advance.
+        /// </summary>
+        [JsonProperty("prebilling")]
+        public List<SubscriptionSchedulePrebillingOptions> Prebilling { get; set; }
+
+        /// <summary>
         /// In cases where the amendment changes the currently active phase, specifies if and how to
         /// prorate at the time of the request.
         /// One of: <c>always_invoice</c>, <c>create_prorations</c>, or <c>none</c>.