+++ src/Stripe.net/Services/Subscriptions/SubscriptionUpdateOptions.cs
@@ -68,6 +68,12 @@ namespace Stripe
         public bool? CancelAtPeriodEnd { get; set; }
 
         /// <summary>
+        /// Details about why this subscription was cancelled.
+        /// </summary>
+        [JsonProperty("cancellation_details")]
+        public SubscriptionCancellationDetailsOptions CancellationDetails { get; set; }
+
+        /// <summary>
         /// Either <c>charge_automatically</c>, or <c>send_invoice</c>. When charging automatically,
         /// Stripe will attempt to pay this subscription at the end of the cycle using the default
         /// source attached to the customer. When sending an invoice, Stripe will email your