+++ src/Stripe.net/Services/Tax/Calculations/CalculationCustomerDetailsOptions.cs
@@ -30,5 +30,14 @@ namespace Stripe.Tax
         /// </summary>
         [JsonProperty("tax_ids")]
         public List<CalculationCustomerDetailsTaxIdOptions> TaxIds { get; set; }
+
+        /// <summary>
+        /// When <c>reverse_charge</c> is provided, the reverse charge rule is applied for taxation.
+        /// When <c>customer_exempt</c> is sent, it treats the customer as tax exempt. Defaults to
+        /// <c>none</c>.
+        /// One of: <c>customer_exempt</c>, <c>none</c>, or <c>reverse_charge</c>.
+        /// </summary>
+        [JsonProperty("taxability_override")]
+        public string TaxabilityOverride { get; set; }
     }
 }